feat: dramatiq & mypy
Squashed commit of the following:

commit c31d13b1043c1b0b59a75fd7adce88407d0a5984
Author: Vinoth Pandian <vinothpandian@users.noreply.github.com>
Date:   Sat Jan 6 00:30:08 2024 -0500

    chore: add mypy & precommit

commit 60e741fd069eec7ff01dea1f644504ac63409697
Author: Vinoth Pandian <vinothpandian@users.noreply.github.com>
Date:   Sat Jan 6 00:02:27 2024 -0500

    feat: add dramatiq

+++ core/redis.py
@@ -11,7 +11,7 @@ CacheKeys = List[str | int]
 
 
 class RedisCache:
-    client = None
+    client: redis.Redis | None = None
 
     def __init__(self, settings: Settings) -> None:
         self.client = redis.Redis(
@@ -32,6 +32,9 @@ class RedisCache:
         return hashlib.md5("".join(map(str, keys)).encode()).hexdigest()
 
     def get(self, keys: CacheKeys) -> np.ndarray | None:
+        if self.client is None:
+            return None
+
         key = self._generate_key(keys)
 
         if data := self.client.get(key):
@@ -41,16 +44,28 @@ class RedisCache:
         return None
 
     def set(self, keys: CacheKeys, value: Any, ex: int = 60 * 5) -> None:
+        if self.client is None:
+            return None
+
         key = self._generate_key(keys)
         data = json.dumps(value)
         self.client.set(key, data, ex=ex)
 
     def delete(self, keys: CacheKeys) -> None:
+        if self.client is None:
+            return None
+
         key = self._generate_key(keys)
         self.client.delete(key)
 
     def ping(self) -> bool:
+        if self.client is None:
+            return False
+
         return self.client.ping()
 
     def disconnect(self) -> None:
+        if self.client is None:
+            return None
+
         self.client.close()
